Massive changes, highly unstable. Relies on a dev version of Composer. Awaiting PR response.

+++ src/sculpin/compiler/Compiler.php
@@ -30,16 +30,30 @@ class Compiler {
     public function compile($pharFile = 'sculpin.phar')
     {
 
+        $process = new Process('git status --porcelain');
+        if ($process->run() != 0) {
+            throw new \RuntimeException('Could not determine clean state from git.');
+        }
+        if (preg_match('/\w/', $process->getOutput())) {
+            // TODO: This should mabe be able to be forced?
+            throw new \RuntimeException('Working directory has changes, refusing to build.');
+        }
+
         if (file_exists($pharFile)) {
             unlink($pharFile);
         }
-
+        
         $process = new Process('git log --pretty="%h" -n1 HEAD');
         if ($process->run() != 0) {
-            throw new \RuntimeException('The git binary cannot be found.');
+            throw new \RuntimeException('Could not determine current version from git.');
         }
         $this->version = trim($process->getOutput());
 
+        $process = new Process('git describe --exact-match HEAD');
+        if ($process->run() == 0) {
+            $this->version = trim($process->getOutput());
+        }
+        
         $phar = new \Phar($pharFile, 0, 'sculpin.phar');
         $phar->setSignatureAlgorithm(\Phar::SHA1);
         
@@ -73,6 +87,7 @@ class Compiler {
             $this->addFile($phar, $file);
         }
         
+        $this->addFile($phar, new \SplFileInfo($this->projectRoot . '/vendor/.composer/ClassLoader.php'));
         $this->addFile($phar, new \SplFileInfo($this->projectRoot . '/vendor/.composer/autoload.php'));
         $this->addFile($phar, new \SplFileInfo($this->projectRoot . '/vendor/.composer/autoload_namespaces.php'));
         
@@ -126,6 +141,8 @@ class Compiler {
  */
 
 Phar::mapPhar('sculpin.phar');
+
+define('SCULPIN_RUNNING_AS_PHAR', true);
 
 require 'phar://sculpin.phar/bin/sculpin';
